Add localStorage
When adding localStorage, Juan Pablo writes this:

const [gastos, setGastos] = useState(
   localStorage.getItem("gastos")
    ? JSON.parse(localStorage.getItem("gastos"))
    : []
);

Which throws an error in React 18.0.0 when mapping in ListadoGastos.

Uncaught TypeError: gastos.map is not a function

That happens, as I can observe, because the compiler can't do first the verification in ? and : and then know that the state 'gastos' is an array.

As localStorage might not exist, therefore the state is undefined and map function does not exist.

I fixed it by declaring the state as an array and then copy with spread operator whatever is in localStorage.

If the localStorage doesn't exists, then it won't copy anything and the state will be an empty array []
But if it exists, then the state will have something [{}, {}, {}]