About Page
So ... it works a bit different than React SPA alone.

It has a router built in with file directory structure, so if you want to go to mydomain.com/about
you would only need to make a file about.js or about.jsx in the same directory than index.js
Something like this:

pages
| api
| _app.js
| index.js
| about.jsx

Note that in React we often write the filename capitalized, like About.jsx when referring to a component.
In Next, that is not the case, because the filename corresponds to the route that will be written on the search bar (forgot the name, the one where you write the web domain, at the top of the browser)

So, if you want to go to mydomain.com/some-page
The component needs to be called some-page.jsx, not Some-page.jsx

Also, the name of the component can have nothing to do with the filename.
So, in about.jsx my component is called AboutPage.