Next Link
So, Next has a Link component that works kind of like React Router Dom v6.

import Link from 'next/link'

<Link href={'/route-to-be-served'}>whatever string goes</Link>

The major change when using Link instead of just anchor tags <a></a> is that it no longer full reloads the page.

If you just use anchor tags, you'll see that it reloads the page (actually, it full reloads, even the dev tools).
That is ... bad, full reloads must be avoided in my opinion.

Link makes it that Next prefetch the component to be rendered in a Link component.
So, in this app, at this stage of development, we only have 2 pages: index.js and about.jsx
Each page has a Link to route to either / or /about

Lets say we are at mydomain.com/ (index.js)
Because we are using Link, when the file index.js is rendered, Next knows it has a Link component that goes to /about, so it prefetches that about page in memory.

The result: a similar behavior as an SPA.

Might be a bit slow if there's way to much Links, honestly idk yet.