Layouts
So ... kind of a new concept.

If we see the code in the index.jsx, about and contact pages, we'll see that there's a LOT of code that is written twice, thrice, etc.

Code that could be in another component, because the only thing that really changes is what's inside the main tag.

Those H1 that says Home Page, About Page, Contact Page.
And those other h1 that has the Link to another page,
the p tag that says to edit whatever file.jsx

Those are the ones that really changes, so ... all the other code is pretty much garbage.

Well, you can use that garbage code in a component that has a special name: a Layout.

Basically, you make a Higher Order Component HOC, that receives a children and renders it.

You guessed it, the children is the code that actually changes in those pages.

And you can have as many layouts as you need, depending on how you want to build the page or how it was designed in UX/UI.


Another thing, there was some changes in the CSS, because of those changes, you can no longer use the styles constant and use them in each index.jsx page.

So, in order to fix the issue that it brings, instead of using className={styles.whatever}, you need to replace it with className = {'whatever'}

It kinda will not work well in Next to simply put className = 'whatever' because of Static Generation ... or so Fernando Herrera says.

Need to re hear what he says on Section 02 - 19. Resolución de tarea sobre Layouts, almost at the end of the video.