Navbar Nextui & 3rd part image handlers
- Added a Navbar component.
It's pretty straight forward, but uses 2 components and 1 hook from Nextui.

The hook: useTheme.
It has a LOT of properties, but here it's only used the property theme.
Theme has to do with colors and other stuffs, so we use theme to make the navbar color a gray darkish color. Only for that.

The components: Text and Spacer.
Text is pretty much a way to render whatever tag it already is a text wrappler. (does that word even exists?)
It goes like <Text color?='some color' h2> Some text here</Text>

Note that I put an H2 tag, you can put Headings, P, Blockquote, etc.

More on this in: https://nextui.org/docs/components/text

Finally, Spacing.
It does that, space. It really converts to a span, but you can customize it. In this example, we used it as a separator from the heading Pokemon, and the button Favoritos, using flex.

- As for the Image component of Next.

If you try and put images from internet, there will be an error.
Basically, because it's trying to do static generation, Next is trying to obtain images sources from the project folder, the public folder i.e.

Internet is, obviously, not a project folder.
So, you need to change configs in next.config.js in order to allow those images to be loaded. It's basically for security purposes, you may point to a location that runs unsecure code.

You must add a domain that you are 100% sure it's safe, or at least that you think it's safe.

Then, you only need to restart yarn dev for the error to go away.