IMPORTANT: getStaticProps
It's a bit hard to get it 1st time.

If you ever forget what this is about, see the video

Section 04: 45. Next - GetStaticProps


So, how to say it. This function only works for Pages, not for basic components.
The function runs in build time, when you yarn build.
This is ahead of time as when some user goes to the page, so the function it's already done by the time it's deployed on internet.

Here you can do a lot of things that are commonly used inside a server, like sending http request, read the filesystem, read databases, etc.
All in order to inject that information to the Page (in this case, only HomePage).
i.e. Get ALL the pokemons from the pokeapi.co, so to be rendered in a grid layout or something.
Or so I think this is going to be used for later on the course.

This, of course, in React you could do it with a useEffect hook, but that will work ONCE the user enters the page, it will be doing it once per user request.
Let's say the API doesn't work some day, you're screwed.
With getStaticProps, the API works fine TODAY, when we build it. Then it will inject the info that an useEffect could do, but generates the pages ahead of time, and it will only do it ONCE. Just for the build time, not for every user request.