Interfaces, API and SSG
The api folder that is in root, IT'S NOT THE API FOLDER from pages. That one is the endpoint on Next.
This api folder on root, is for axios.

Too many changes, forgot to commit on interfaces.

Installed axios too.

So, we began by request to pokeapi 151 pokemons.
Each request, stored in data, is of type PokemonListResponse, which we created an interface with Postman and PasteJSONasCode.

That interface has some errors by default, mainly previous, the extension thinks previous is always null, which is not. So, previous?: string;
Also, renamed results type as SmallPokemon[]

SmallPokemon has, by default, only 2 properties: name and url.
We added id and img, because we wanted to have an id (later on used within li key) and a svg image path.

Those properties that does not come with the axios response, we added them manually with the map function that already knows that it is an array because of the interface.

Once you have the new array, you pass them as props to HomePage, and there you can also use the map function (again, because TS knows it's an array) and render 151 pokemons.

When you make a yarn build, it generates a JSON with all those objects: {name, url, id, img} so when you use yarn start, you'll no longer request anything to pokeapi.co each time you load the page, because it already did it for you with staticProps.

That is called Static Site Generation (SSG).
That's the magic about it.