Nextui Cards
So it works a bit like bootstrap, sort of.

It's straight forward.
Instead of using ul and li, we use the cards.

Also, according to Fernando Herrera, it implements lazy loading, so the cards will be rendered once it knows it's about to enter the viewport.
I don't know how to actually assert that statement, but ... well, I guess we believe Nextui devs or Fernando Herrera.

Also, he uses in the video the properties hoverable and clickeable.
By the time I'm doing the course, those properties changed their names to isHoverable and isPressable.

More on that in:
https://nextui.org/docs/components/card