Moved the grid to a PokemonCard component
For ease of readability, we move all the content that the map function does in an independent component.

There was a way a bit hard to implement the props on that component.

I could have used this:

export const PokemonCard: FC<Props> = ( { pokemon: {id, img, url, name} } ) => {...}

Its a bit hard to read and understand, but it works.

Instead I use what I used, simply destructure the pokemon in a const and use those.