IMPORTANT: getStaticPaths
Woooh, there's a lot of things to understand here.

Again, see the video Section 04: 52. Next - getStaticPaths and 53. Generando las 151 páginas de Pokémons

Where to even begin?

You write the getStaticPaths function before the getStaticProps. Not sure if it NEEDS to be before, but anyway.

GetstaticPaths is the one in charge to create ALL 151 PAGES when yarn build.

Later on, getStaticPaths allows StaticProps to obtain the params on the url, in this case, just the id as a string, using the context.
The context has way more properties, but we destructure it to obtain only params.

From the params, we obtain the id, and with that ID we axios the shit out of the pokemon ID.

We send that data as props to the page and there you can do whatever the hell you want with it.

In dev mode it is slow af, but when you yarn build, the build process is the one and only time it will be slow, because it is fetching the data and stuff from all 151 pokemons.

Once it finishes, yarn start will serve all the pages and it will be smooth fast... tho, not as fast as simple React App with vite or CRA.