NextLink & NextUILink
So, we pass the href property to Link (NextUI) so it can have the Href.

Basically, NextLink will push to a page with the href='some/page' but if you have more than 1 children, then it throws an error.
If you try and just put a fragment there, then the error will go away BUT it will not work, because there's no anchor tag.
The anchor tag comes with Link component (NextUI), and therefore for this anchor tag to obtain the href, you need to pass the property passHref to NextLink component.