05 - localStorage & favorites
In Section 05 - 62. Guardar en localStorage, Fernando mentions that although you can use React as normal, there are some places where some functions are not defined for the server.

i.e localStorage. If you clg localStorage.getItem(...), outside of a function, it will crash, because when the component is rendered in the server, it will try to call the localStorage function and because localStorage is not defined in server, then it crashes.

If you put it in a function, it will sort of work. In the web client console, it works, and it is defined, but in the console of the server then somethings are still not defined.

Thats kind of a way to know wheter or not something is running in the server or in the client.