[ADD] ISG and ISR
Incremental Static Generation and Incremental Static Regeneration.

Weird stuff, a bit ambiguous the way Fernando explained ISR.

Should see: https://nextjs.org/docs/basic-features/data-fetching/incremental-static-regeneration

Basically, added functionality to create new pages by ID or name that were not in the first build.
The first build only had 151 pokemons, but if you look for pokemon 152 / chikorita, it would simply put a 404 page.
With ISG and ISR (?) we would not push a 404 page, but the pokemon page instead.
If the ID or name doesn't exists, it pushes you to homepage. But if it exists, it will create that page.

I believe it will actually create the page in vercel too. Let's find out.